--- app/src/main/java/dev/cancio/filmin/MyApplication.kt
@@ -1,10 +1,0 @@
-package dev.cancio.filmin
-
-import android.app.Application
-import dev.cancio.filmin.di.DaggerMovieComponent
-import dev.cancio.filmin.di.MovieComponent
-
-class MyApplication: Application() {
-
-    val appComponent = DaggerMovieComponent.builder().build()
-}
