--- app/src/main/java/dev/cancio/filmin/di/MovieComponent.kt
@@ -1,11 +1,0 @@
-package dev.cancio.filmin.di
-
-import dagger.Component
-import dev.cancio.filmin.ui.activity.MainActivity
-
-@Component(
-    modules = [MovieModule::class]
-)
-interface MovieComponent {
-    fun inject(activity: MainActivity)
-}
