--- app/src/main/java/dev/cancio/filmin/di/MovieModule.kt
@@ -1,14 +1,0 @@
-package dev.cancio.filmin.di
-
-import dagger.Module
-import dagger.Provides
-import dev.cancio.filmin.data.repository.MovieRepository
-
-@Module
-class MovieModule {
-
-    @Provides
-    fun provideMovie(): MovieRepository {
-        return MovieRepository()
-    }
-}
