+++ app/src/main/java/dev/cancio/popcorn/di/MovieModule.kt
@@ -1,0 +1,14 @@
+package dev.cancio.popcorn.di
+
+import dagger.Module
+import dagger.Provides
+import dev.cancio.popcorn.data.repository.MovieRepository
+
+@Module
+class MovieModule {
+
+    @Provides
+    fun provideMovie(): MovieRepository {
+        return MovieRepository()
+    }
+}
