+++ app/src/main/java/dev/cancio/popcorn/ui/activity/NoneActivity.kt
@@ -1,0 +1,12 @@
+package dev.cancio.popcorn.ui.activity
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import dev.cancio.popcorn.R
+
+class NoneActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_none)
+    }
+}
