+++ app/src/main/res/menu/bottom_bar_menu.xml
@@ -14,7 +14,7 @@
     <item
         android:id="@+id/globe_item"
         android:title="@string/bottom_bar_item_globe"
-        app:showAsAction="ifRoom"
+        app:showAsAction="always"
         android:icon="@drawable/ic_globe"/>
     <item
         android:id="@+id/profile_item"
