+++ app/src/main/res/values/colors.xml
@@ -10,5 +10,6 @@
     <color name="grey_2">  #FF8F908F</color>
     <color name="black"> #FF000000</color>
     <color name="white">  #FFFFFFFF</color>
+    <color name="transparent">  #00000000</color>
 
 </resources>
